Restructured to a unified binary for client and server
