--- client/go.mod
@@ -1,10 +1,0 @@
-module github.com/nugget/vanitykeygen/client
-
-go 1.23.4
-
-require (
-	github.com/mikesmitty/edkey v0.0.0-20170222072505-3356ea4e686a
-	golang.org/x/crypto v0.31.0
-)
-
-require golang.org/x/sys v0.28.0 // indirect
