+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/nugget/vanitykeygen
+
+go 1.23.4
+
+require (
+	github.com/mikesmitty/edkey v0.0.0-20170222072505-3356ea4e686a
+	golang.org/x/crypto v0.32.0
+)
+
+require golang.org/x/sys v0.29.0 // indirect
