+++ pkg/server/.gitignore
@@ -1,0 +1,2 @@
+server
+matchfile.log
