--- server/.gitignore
@@ -1,2 +1,0 @@
-server
-matchfile.log
