--- server/go.mod
@@ -1,3 +1,0 @@
-module github.com/nugget/vanitykeygen/server
-
-go 1.23.4
